new: entirely new shift system
Yeah, I know I said I wasn’t going to do this, but this code was added to an older game at their request for the Changing Breeds sourcebook. It also gave me a chance to create a system that would be usable in the newer WoD/CoD code as well, so here we are.